requirements.txt
ollama>=0.6.0
openai-whisper>=20250625
pyttsx3>=2.99
sounddevice>=0.5.0
numpy>=1.26.0
